requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
pydantic>=2.0.0
